+++ cmd/band/band.go
@@ -20,5 +20,6 @@ func BandCommandNew(event *event.Event) *cobra.Command {
 		},
 	}
 	bandCommand.AddCommand(BandCommandCreateNew(event))
+	bandCommand.AddCommand(BandCommandListNew(event))
 	return bandCommand
 }